requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-allauth-ui==0.2.2
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-filter==24.1
django-htmx==1.17.3
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-richtextfield==1.6.2
django-shortuuidfield==0.1.3
django-widget-tweaks==1.5.0
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
whitenoise==6.6.0
dj-database-url==2.1.0
django-environ==0.11.2